+++ src/core/login/splitting.ts
@@ -3,7 +3,6 @@ import { base64 } from 'rfc4648'
 
 import {
   EdgeCurrencyWallet,
-  EdgeMetadata,
   EdgeSpendInfo,
   EdgeWalletInfo,
   EdgeWalletStates
@@ -180,9 +179,11 @@ export async function splitWalletInfo(
 }
 
 async function protectBchWallet(wallet: EdgeCurrencyWallet): Promise<void> {
+  const bchCurrency = { currencyCode: 'BCH', tokenId: null }
+
   // Create a UTXO which can be spend only on the ABC network
   const spendInfoSplit: EdgeSpendInfo = {
-    currencyCode: 'BCH',
+    ...bchCurrency,
     spendTargets: [
       {
         nativeAmount: '10000',
@@ -199,12 +200,16 @@ async function protectBchWallet(wallet: EdgeCurrencyWallet): Promise<void> {
   await wallet.saveTx(broadcastedSplitTx)
 
   // Taint the rest of the wallet using the UTXO from before
-  const { publicAddress } = await wallet.getReceiveAddress()
+  const { publicAddress } = await wallet.getReceiveAddress(bchCurrency)
   const spendInfoTaint: EdgeSpendInfo = {
-    currencyCode: 'BCH',
-    spendTargets: [{ publicAddress, nativeAmount: '0' }],
-    metadata: {},
-    networkFeeOption: 'high'
+    ...bchCurrency,
+    metadata: {
+      name: 'Replay Protection Tx',
+      notes:
+        'This transaction is to protect your BCH wallet from unintentionally spending BSV funds. Please wait for the transaction to confirm before making additional transactions using this BCH wallet.'
+    },
+    networkFeeOption: 'high',
+    spendTargets: [{ publicAddress, nativeAmount: '0' }]
   }
   const maxAmount = await wallet.getMaxSpendable(spendInfoTaint)
   spendInfoTaint.spendTargets[0].nativeAmount = maxAmount
@@ -212,12 +217,6 @@ async function protectBchWallet(wallet: EdgeCurrencyWallet): Promise<void> {
   const signedTaintTx = await wallet.signTx(taintTx)
   const broadcastedTaintTx = await wallet.broadcastTx(signedTaintTx)
   await wallet.saveTx(broadcastedTaintTx)
-  const edgeMetadata: EdgeMetadata = {
-    name: 'Replay Protection Tx',
-    notes:
-      'This transaction is to protect your BCH wallet from unintentionally spending BSV funds. Please wait for the transaction to confirm before making additional transactions using this BCH wallet.'
-  }
-  await wallet.saveTxMetadata(broadcastedTaintTx.txid, 'BCH', edgeMetadata)
 }
 
 /**